Implement first reconciler bits (#22)
* Add initial reconciler package

Adds the reconciler package, with a basic skeletal non-functional
reconciler. Includes Ginkgo for testing, Gomega for matchers for Ginkgo,
and some other imports required by those two.

Signed-off-by: Drew Erny <drew.erny@docker.com>

* Add Reconciler specs

Adds the ginkgo specs for the reconciler object. Also changes the
reconciler to use interfaces.BackendClient

Signed-off-by: Drew Erny <drew.erny@docker.com>

* Write first reconciler tests

* Finish ReconcileStack for services

Finishes the implementation of ReconcileStack, but only for the parts
that handle services.

Signed-off-by: Drew Erny <drew.erny@docker.com>

* Simplify Reconciler Interface

It doesn't really need to have different entry points for stacks and
other resource types. I don't know why I wrote it like that in the first
place.

Signed-off-by: Drew Erny <drew.erny@docker.com>