More reconciler work (#27)
* Add Dispatcher and Manager

Adds the Dispatcher and Manager components of the Reconciler. The
Dispatcher handles what objects the Reconciler should process and when,
and the Manager is the entrypoint for the whole package.

Signed-off-by: Drew Erny <drew.erny@docker.com>

* Break reconciler components into more packages

Moves the various reconciler sub-components into their own packages,
encouraging an even greater degree of decoupling.

Signed-off-by: Drew Erny <drew.erny@docker.com>

* Add support for reconciling services

Adds support for reconciling services. Handles cases where the service
needs updates, the service needs no updates, and the service needs to be
deleted.

Signed-off-by: Drew Erny <drew.erny@docker.com>

* Begin work on dispatcher

Starts work on the dispatcher. Implements the basic state machine and a
test.

Signed-off-by: Drew Erny <drew.erny@docker.com>

* fix review comments

Signed-off-by: Drew Erny <drew.erny@docker.com>

* add exception for reconciler mock and reindent file

Signed-off-by: Drew Erny <drew.erny@docker.com>