Check in generated schema data for downstream
Without checking this generated file in, downstream
projects that need to vendor wont be able to build
or will have to generate on their own which gets messy.

This will automatically update it if the content has changed, so it'll
leave your local tree dirty to help us remember to commit updates
when needed.

Signed-off-by: Daniel Hiltgen <daniel.hiltgen@docker.com>