Add desired/running tasks to status (#39)
The current CLI (and GUI) implements a complex client-side algorithm
to attempt to determine the desired number of tasks and current
running tasks.  The current running count is ~easy to determine for
the clients, but the desired count requires deep understanding of the
scheduling parameters for the given service, which in essence requires
re-implementing the full scheduler algorithm client side - not a good
pattern.  This moves it server side so we can keep the algorithms aligned.

Signed-off-by: Daniel Hiltgen <daniel.hiltgen@docker.com>